всё работает

+++ src/Drago/Coordinates.java
@@ -11,7 +11,7 @@ public class Coordinates {
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
